+++ PrefeFortune/Views/Components/FortuneButton.swift
@@ -20,22 +20,8 @@ struct FortuneButton: View {
                 }
             } label: {
                 Text("占う")
-                    .font(.headline)
-                    .foregroundColor(.white)
-                    .padding(.vertical, 25)
-                    .frame(maxWidth: 200)
-                    .frame(minWidth: 100)
-                    .background(
-                        RoundedRectangle(cornerRadius: 20)
-                            .fill(Color.customYellowColor)
-                            .shadow(color: .black.opacity(0.3), radius: 10, x: 0, y: 5)
-                            .overlay(
-                                RoundedRectangle(cornerRadius: 20)
-                                    .stroke(Color.white.opacity(0.9), lineWidth: 1)
-                            )
-                    )
-                    .padding(.horizontal, 30)
             }
+            .ButtonStyleModifier()
             .disabled(!isFormComplete)
             .opacity(isFormComplete ? 0.8 : 0.5)
             Spacer()
