+++ PrefeFortune/Views/StartNavigation/StartNavigatinoView.swift
@@ -79,16 +79,8 @@ struct StartNavigationView: View {
             }
         }) {
             Text(buttonText)
-                .padding()
-                .font(.title2)
-                .fontWeight(.bold)
-                .frame(width: 200, height: 100)
-                .background(Color.customYellowColor)
-                .foregroundColor(.white)
-                .cornerRadius(50)
-                .opacity(0.8)
-                .shadow(color: .black.opacity(0.3), radius: 10, x: 3, y: 3)
         }
+        .ButtonStyleModifier()
     }
 
     private func handleButtonTap() {
